requirements.txt
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
